move the related code to CharmInfo in this own file
*original idea of Ovahlord
